+++ Frontend/my-app/.dockerignore
@@ -1,0 +1,6 @@
+.git
+.gitignore
+.dockerignore
+node_modules
+npm-debug.log
+README.md
