+++ Frontend/my-app/Dockerfile
@@ -1,0 +1,35 @@
+# Stage 1: Install dependencies
+FROM node:18-slim AS deps
+WORKDIR /app
+COPY package.json package-lock.json ./
+RUN npm install
+
+# Stage 2: Build the application
+FROM node:18-slim AS builder
+WORKDIR /app
+COPY --from=deps /app/node_modules ./node_modules
+COPY . .
+RUN npm run build
+
+# Stage 3: Production image
+FROM node:18-slim AS runner
+WORKDIR /app
+
+# Create a non-root user for security
+RUN addgroup --system --gid 1001 nodejs
+RUN adduser --system --uid 1001 --ingroup nodejs --no-create-home nextjs
+
+# Copy built assets from the builder stage
+COPY --from=builder /app/public ./public
+COPY --from=builder --chown=nextjs:nodejs /app/.next ./.next
+COPY --from=builder /app/node_modules ./node_modules
+COPY --from=builder /app/package.json ./package.json
+
+USER nextjs
+
+EXPOSE 3000
+
+ENV PORT 3000
+
+# Start the Next.js app
+CMD ["npm", "start"]
