+++ Frontend/my-app/src/components/logo.png
