+++ backend/requirements.txt
@@ -1,2 +1,4 @@
 django
 gunicorn
+djangorestframework==3.14.0  
+django-cors-headers==4.2.0
