--- frontend/.dockerignore
@@ -1,17 +1,0 @@
-node_modules
-npm-debug.log
-Dockerfile
-.dockerignore
-.git
-.gitignore
-README.md
-.env
-.nyc_output
-coverage
-.coverage
-.coverage.*
-.cache
-.DS_Store
-*.log
-build
-dist
