--- frontend/Dockerfile
@@ -1,45 +1,0 @@
-# Build stage
-
-FROM node:18-alpine AS builder
-
-
-
-WORKDIR /app
-
-
-
-COPY package.json package-lock.json* ./
-
-RUN npm install
-
-
-
-COPY . .
-
-RUN npm run build
-
-
-
-# Production stage
-
-FROM nginx:alpine
-
-
-
-# Copy built assets from builder
-
-COPY --from=builder /app/build /usr/share/nginx/html
-
-
-
-# Copy custom nginx config (optional)
-
-# COPY nginx.conf /etc/nginx/nginx.conf
-
-
-
-EXPOSE 80
-
-
-
-CMD ["nginx", "-g", "daemon off;"]
