--- frontend/src/App.css
@@ -1,178 +1,0 @@
-/* General Body & App Container Styles */
-body {
-    margin: 0;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-      sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-color: #f0f2f5;
-  }
-  
-  .app-container {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  /* Header Component */
-  .header {
-    background-color: #ffffff;
-    border-bottom: 1px solid #e0e0e0;
-    padding: 0 2rem;
-  }
-  
-  .header-content {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    max-width: 1200px;
-    margin: 0 auto;
-    height: 80px;
-  }
-  
-  .logo {
-    height: 50px;
-  }
-  
-  .main-nav a {
-    margin-left: 20px;
-    text-decoration: none;
-    color: #333;
-    font-weight: 500;
-    font-size: 1rem;
-  }
-  
-  /* SubHeader Component */
-  .sub-header {
-    background-color: #6a7f4b;
-    color: #ffffff;
-    padding: 0 2rem;
-  }
-  
-  .sub-header-content {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    gap: 20px;
-    max-width: 1200px;
-    margin: 0 auto;
-    height: 40px;
-  }
-  
-  .org-selector {
-    background-color: transparent;
-    color: #ffffff;
-    border: none;
-    font-size: 1rem;
-    cursor: pointer;
-    -webkit-appearance: none;
-    -moz-appearance: none;
-    appearance: none;
-    padding-right: 1em; /* space for the arrow */
-  }
-  
-  .org-selector option {
-    background-color: #333;
-  }
-  
-  /* Main Content & Dashboard Layout */
-  .main-content {
-    padding: 2rem;
-  }
-  
-  .dashboard-layout {
-    display: grid;
-    grid-template-columns: 300px 1fr;
-    gap: 2rem;
-    max-width: 1200px;
-    margin: 0 auto;
-  }
-  
-  /* Card Component (Used by MyAccount and Organizations) */
-  .card {
-    background-color: #ffffff;
-    border: 1px solid #e0e0e0;
-    border-radius: 8px;
-    padding: 1.5rem;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-    height: fit-content;
-  }
-  
-  .card-title {
-    font-size: 1.1rem;
-    font-weight: 500;
-    color: #333;
-    margin-top: 0;
-    margin-bottom: 1.5rem;
-    padding-bottom: 0.75rem;
-    border-bottom: 1px solid #eee;
-  }
-  
-  /* MyAccount Component Specifics */
-  .account-details p {
-    margin: 0 0 0.5rem 0;
-    color: #555;
-    font-size: 0.9rem;
-  }
-  
-  .account-details strong {
-    color: #000;
-  }
-  
-  .account-links {
-    margin-top: 1rem;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .account-links a {
-    color: #318a22;
-    text-decoration: none;
-    margin-bottom: 0.5rem;
-    font-size: 0.9rem;
-  }
-  
-  .account-links a:hover {
-    text-decoration: underline;
-  }
-  
-  /* Organizations Component Specifics */
-  .org-actions {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .action-link {
-    background-color: #f7f7f7;
-    border: 1px solid #ddd;
-    padding: 1rem;
-    margin-bottom: 1rem;
-    text-decoration: none;
-    color: #333;
-    border-radius: 4px;
-    text-align: left;
-    transition: background-color 0.2s;
-  }
-  
-  .action-link:hover {
-    background-color: #efefef;
-  }
-  
-  .action-link:last-child {
-    margin-bottom: 0;
-  }
-  
-  /* Responsive adjustments */
-  @media (max-width: 768px) {
-    .dashboard-layout {
-      grid-template-columns: 1fr;
-    }
-    .header-content {
-      flex-direction: column;
-      height: auto;
-      padding: 1rem 0;
-    }
-    .main-nav {
-      margin-top: 1rem;
-    }
-  }
